refactor: rename api entrypoint
