+++ opentelemetry-kotlin-implementation/src/commonMain/kotlin/io/embrace/opentelemetry/kotlin/OpenTelemetryInstanceImpl.kt
@@ -1,6 +1,9 @@
 package io.embrace.opentelemetry.kotlin
 
 import io.embrace.opentelemetry.kotlin.clock.ClockImpl
+import io.embrace.opentelemetry.kotlin.creator.ObjectCreator
+import io.embrace.opentelemetry.kotlin.init.LoggerProviderConfigDsl
+import io.embrace.opentelemetry.kotlin.init.TracerProviderConfigDsl
 
 /**
  * Constructs an [OpenTelemetry] instance that uses the opentelemetry-kotlin implementation.
@@ -8,5 +11,8 @@ import io.embrace.opentelemetry.kotlin.clock.ClockImpl
 @ExperimentalApi
 @Suppress("UNUSED_PARAMETER")
 public fun OpenTelemetryInstance.default(
-    clock: Clock = ClockImpl()
+    tracerProvider: TracerProviderConfigDsl.() -> Unit = {},
+    loggerProvider: LoggerProviderConfigDsl.() -> Unit = {},
+    clock: Clock = ClockImpl(),
+    objectCreator: ObjectCreator = throw UnsupportedOperationException()
 ): OpenTelemetry = throw UnsupportedOperationException()
